Fix TextWildcards not disabling dynamicPrompts
ComfyUI has built-in dynamic prompts now, though they only support
simple {a|b} and not the more advanced {2$$-$$a|b|c} dynamic prompts
with nesting. In addition built-in dynamic prompts would overwrite the
prompt in the field when trying to load generation info. By setting
"dynamicPrompts" to False, we disable that functionality so this node
can do it's job properly again.